requirements.txt
Django
djangorestframework
celery 
redis
requests-oauthlib
django-celery-beat
django-redis
drf-spectacular